+++ data-raw/df data.R
@@ -246,4 +246,16 @@ hardlyworking <- data.frame(
 hardlyworking$is_senior <- hardlyworking$seniority > 4
 
 save(hardlyworking, file = "data/hardlyworking.rdata")
+
+
+# rouder2016 --------------------------------------------------------------
+
+rouder2016 <- read.table("data-raw/effectSizePuzzler.txt",
+  header = TRUE, stringsAsFactors = FALSE
+)
+rouder2016[["id"]] <- factor(rouder2016[["id"]])
+rouder2016[["cond"]] <- factor(rouder2016[["cond"]])
+
+save(rouder2016, file = "data/rouder2016.rdata")
+
 # styler: on